Closes #22 exception handling added

+++ in28minutes/src/main/java/com/in28minutes/springboot/web/controller/TodoController.java
@@ -47,6 +47,9 @@ public class TodoController {
 
 	@RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam int id) {
+		if(id == 1) {
+			throw new RuntimeException("Something went wrong");
+		}
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
